.groups argument in dplyr:: summarise
dplyr 1.0.0 provides a convenient option, `.groups`, for `summarise` to keep or drop the grouping.  In the example, we use `group_by(village, memb_assoc)` before calculating the `mean(no_membrs)`, which groups the dataframe by village.  Before dplyr 1.0.0, we need to use `ungroup` to ungroup the dataframe.  Now if we specify `.groups = "drop"` in `summarise`, the grouping will be dropped after the mean being calculated.   Therefore, to ungroup, we use

```{r}
interviews_items_owned %>%
    mutate(number_items = rowSums(select(., bicycle:television))) %>%
    group_by(village) %>%
    summarize(mean_items = mean(number_items), .groups = "drop")
```